+++ public/js/app.js
@@ -8,13 +8,15 @@ angular.module('mean', ['ngCookies', 'ngResource', 'ui.router', 'ui.bootstrap', 
         oauthType: '1.0',
         popupOptions: { width: 495, height: 645 }
     });
+
     $authProvider.google({
-        clientId: '133157744382-khfvhevgh114s5bsba5pbql6lu51l9ad.apps.googleusercontent.com',
+        clientId: 'your google client id here', // google client id
         url: '/auth/google',
         redirectUri: 'http://localhost:3000/auth/google/callback'
-
     });
+
 });
+
 angular.module('mean.system', []);
 angular.module('mean.articles', []);
 angular.module('mean.auth', []);
